+++ docs/changes.rst
@@ -8,7 +8,8 @@ To be released.
 
 - Added :class:`sider.types.Date` type.
 - Added :class:`sider.datetime.FixedOffset` tzinfo subtype.
-- Added :class:`sider.types.DateTime` type.
+- Added :class:`sider.types.DateTime` and
+  :class:`~sider.types.TZDatetime` types.
 
 
 Version 0.1.0